refactor: Polynomial 클래스에서 사용되던 Collection 관련 메서드들을 CollectionUtils 클래스로 추출
CollectionUtils 제공 기능:
1. convertListToStack: 리스트를 스택으로 변환
2. reverse: 입력 리스트의 순서를 뒤집어 새로운 리스트 생성
