+++ src/test/java/calculator/BinaryOperatorTest.java
@@ -1,24 +1,12 @@
 package calculator;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
 
 import static org.assertj.core.api.Assertions.*;
 
 public class BinaryOperatorTest {
-    @BeforeEach
-    void setUp() {
-    }
-
-    @AfterEach
-    void tearDown() {
-
-    }
-
     @ParameterizedTest
     @CsvSource({"10, 120"
             , "10, -123"
